refactored comoponent directories
